server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pyjwt
bcrypt
pytesseract
Pillow
langchain
langchain-huggingface
langchain-community
google-genai
requests
websockets
